@@ -7,17 +7,21 @@ import { AppComponent } from './app.component';
 import { CalculatorComponent } from './calculator/calculator.component';
 import {ProfitcalculatorService} from './services/profitcalculator.service';
 import { DonateComponent } from './donate/donate.component';
+import {MaterialModule} from './material/material.module';
+import { SidenavComponent } from './sidenav/sidenav.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     CalculatorComponent,
-    DonateComponent
+    DonateComponent,
+    SidenavComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpClientModule
+    HttpClientModule,
+    MaterialModule
   ],
   providers: [ProfitcalculatorService],
   bootstrap: [AppComponent]
